Added handling for empty timestamps
When timestamps were not saved in the file metadata, getTimestamps
would return a "reference to non-struct array" error. I added some
statements which would check that the call to getPlaneDeltaT is not
empty. If it is, the code will return a warning along with a timestamp
vector which is all zeros.