Fixed issue with missing tsunits
The last commit addressed an issue when a file was missing timestamp
information. The code will now output a warning + a timestamp vector of
zeros. However, the timestamp units were not being assigned in this
case.

In this commit, tsunits are now assigned to a blank string ''. This
should avoid issues with code compatibility when the timestamp
information is missing.